requirements.txt
Django
djangorestframework
Pillow
psycopg2-binary
python-decouple
drf-yasg
django-filter
djangorestframework-simplejwt
